nopti_back/requirements.txt
fastapi
uvicorn
GoogleNews
aiohttp
beautifulsoup4
OpenAI
python-dotenv


google-api-python-client
google-auth-httplib2
google-auth-oauthlib


elevenlabs

argparse
langchain
langchain_openai
youtube_transcript_api

python-multipart

spotipy
llama-stack
"llama-cpp-python[server]"
llama.cpp
sqlalchemy
elevenlabs